Add option for ng build folder

+++ waxe_image/scripts/get_ng_build.py
@@ -2,6 +2,7 @@ import StringIO
 import os
 import requests
 import shutil
+import sys
 import zipfile
 
 from waxe_image import __version__
@@ -11,9 +12,18 @@ API_RELEASES_URL = 'https://api.github.com/repos/waxe/waxe-image/releases'
 NG_BUILD_FOLDER = 'website'
 
 
-def main():
+def main(argv=sys.argv):
+    if len(argv) > 2:
+        print 'Too many arguments'
+        sys.exit(1)
+    global NG_BUILD_FOLDER
+    if len(argv) == 2:
+        NG_BUILD_FOLDER = argv[1]
     if os.path.isdir(NG_BUILD_FOLDER):
         shutil.rmtree(NG_BUILD_FOLDER)
+    if os.path.exists(NG_BUILD_FOLDER):
+        print 'There is an issue with the folder %s' % NG_BUILD_FOLDER
+        sys.exit(1)
     r = requests.get(API_RELEASES_URL)
     if r.status_code != 200:
         raise ValueError('Bad status code %s' % r.status_code)
